Move StandardProcessRunReporter and JobQuery to job package
