+++ java/clients/src/main/java/sleeper/status/report/ingest/job/IngestJobStatusReportArguments.java
@@ -17,8 +17,8 @@
 package sleeper.status.report.ingest.job;
 
 import sleeper.console.ConsoleInput;
-import sleeper.status.report.query.JobQuery;
-import sleeper.status.report.query.JobQueryArgument;
+import sleeper.status.report.job.query.JobQuery;
+import sleeper.status.report.job.query.JobQueryArgument;
 
 import java.io.PrintStream;
 import java.time.Clock;
