+++ java/clients/src/main/java/sleeper/status/report/ingest/job/IngestJobStatusReporter.java
@@ -17,7 +17,7 @@
 package sleeper.status.report.ingest.job;
 
 import sleeper.ingest.job.status.IngestJobStatus;
-import sleeper.status.report.query.JobQuery;
+import sleeper.status.report.job.query.JobQuery;
 
 import java.util.List;
