+++ java/clients/src/main/java/sleeper/status/report/ingest/job/JsonIngestJobStatusReporter.java
@@ -23,7 +23,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializer;
 import sleeper.ingest.job.status.IngestJobStatus;
-import sleeper.status.report.query.JobQuery;
+import sleeper.status.report.job.query.JobQuery;
 
 import java.io.PrintStream;
 import java.time.Instant;
