+++ java/clients/src/main/java/sleeper/status/report/job/query/JobQueryArgument.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.status.report.query;
+package sleeper.status.report.job.query;
 
 import java.util.HashMap;
 import java.util.Map;
