+++ java/clients/src/test/java/sleeper/status/report/compaction/job/CompactionJobStatusReportArgumentsTest.java
@@ -18,7 +18,7 @@ package sleeper.status.report.compaction.job;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.status.report.query.JobQuery.Type;
+import static sleeper.status.report.job.query.JobQuery.Type;
 
 public class CompactionJobStatusReportArgumentsTest {
