+++ java/clients/src/test/java/sleeper/status/report/query/JobQueryTest.java
@@ -26,7 +26,7 @@ import java.util.Optional;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.when;
-import static sleeper.status.report.query.JobQuery.Type;
+import static sleeper.status.report.job.query.JobQuery.Type;
 
 public class JobQueryTest extends JobQueryTestBase {
     @Test
